+++ jubeatools/formats/jubeat_analyser/memo1/dump.py
@@ -248,7 +248,11 @@ def _dump_memo1_internal(song: Song, circle_free: bool) -> List[ChartFile]:
         timing = chart.timing or song.global_timing
         assert timing is not None
         contents = _dump_memo1_chart(
-            difficulty, chart, song.metadata, timing, circle_free,
+            difficulty,
+            chart,
+            song.metadata,
+            timing,
+            circle_free,
         )
         files.append(ChartFile(contents, song, difficulty, chart))