+++ jubeatools/formats/jubeat_analyser/memo1/load.py
@@ -76,6 +76,9 @@ class Memo1LoadedSection:
 
 
 class Memo1Parser(JubeatAnalyserParser):
+
+    FORMAT_TAG = "#memo1"
+
     def __init__(self) -> None:
         super().__init__()
         self.current_chart_lines: List[DoubleColumnChartLine] = []
@@ -146,6 +149,7 @@ class Memo1Parser(JubeatAnalyserParser):
 
     def load_line(self, raw_line: str) -> None:
         line = raw_line.strip()
+        self.raise_if_separator(line, self.FORMAT_TAG)
         if is_command(line):
             command, value = parse_command(line)
             self.handle_command(command, value)
@@ -235,7 +239,8 @@ class Memo1Parser(JubeatAnalyserParser):
             )
             if arrow_to_note_candidates:
                 solution = pick_correct_long_note_candidates(
-                    arrow_to_note_candidates, frame.position_part,
+                    arrow_to_note_candidates,
+                    frame.position_part,
                 )
                 for arrow_pos, note_pos in solution.items():
                     should_skip.add(arrow_pos)